feat: add initial p2tr output script support
This introduces support for creating p2tr output scripts and addresses
from one or more pubkeys and/or one or more scripts according to BIPs
340 & 341.

Multiple pubkeys are aggregated according to the procedure defined by
MuSig whereby all keys are concatenated and hashed, with the resulting
hash multiplied against each pubkey to tweak it. These tweaked pub keys
are added together to produce the internal pubkey. If only one pubkey
is provided, then the internal pubkey is simply set to that. And if no
pubkeys are provided, a key with no known secret key is used instead.

Any scripts provided are arranged into a Huffman tree according to their
weighted probability and then hashed to form a merkle tree. The root
hash of this tree is then added to the internal pubkey to tweak it. This
tweaked key is finally used for creating the output script which is then
encoded using Bech32m.

Test cases in this commit include examples from the Bitcoin Optech
taproot workshop that derive a Bech32m address from scripts and an
internal key.

Follow-ups to this commit should include expanding the coverage and
number of test cases as well as possibly restructuring existing code or
moving logic to other packages/modules. We should also consider support
for taproot output descriptors, which are currently under review and
developmnet in Bitcoin Core.

Ticket: BG-35571