feat: add MuSig2* key agg & p2tr output support
This introduces support for creating p2tr output scripts and addresses
from one or more pubkeys and/or one or more scripts according to BIPs
340 & 341.

Multiple pubkeys are aggregated according to the MuSig2* procedure
currently being proposed in secp256k1-zkip, whereby all keys are
concatenated  and hashed, with the resulting hash multiplied against
each pubkey to tweak it. MuSig2*, compared to earlier versions of MuSig,
skips the key exponentiation step for the second key. These tweaked pub
keys are added together to produce the internal pubkey. If only one
pubkey is provided, then the internal pubkey is simply set to that. And
if no pubkeys are provided, a key with no known secret key is used
instead.

Any scripts provided are arranged into a Huffman tree according to their
weighted probability and then hashed to form a merkle tree. The root
hash of this tree is then added to the internal pubkey to tweak it. This
tweaked key is finally used for creating the output script which is then
encoded using Bech32m.

Test cases in this commit are derived from secp256k1-zkip examples for
MuSig2* key aggregation and from the Bitcoin Optech taproot workshop
for TapTree construction and for using the TapTree root to derive a
Bech32m address from scripts and an internal key.

Ticket: BG-35571